Removed the unnecesary BaseIntegrationTestsConfig interface, moved the IModelsClientsTestsConfig class into common utils to avoid copypaste
