+++ tests/imodels-clients-tests/src/integration/common/TestDiContainerProvider.ts
@@ -2,10 +2,9 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
+import * as path from "path";
 import { Container } from "inversify";
-import { BaseIntegrationTestsConfig, TestUtilBootstrapper, TestUtilTypes } from "@itwin/imodels-client-test-utils";
-import { IModelsClientsTestsConfig } from "./IModelsClientsTestsConfig";
-import "reflect-metadata";
+import { TestUtilBootstrapper } from "@itwin/imodels-client-test-utils";
 
 let container: Container;
 export function getTestDIContainer(): Container {
@@ -13,7 +12,7 @@ export function getTestDIContainer(): Container {
     return container;
 
   container = new Container();
-  TestUtilBootstrapper.bind(container);
-  container.bind<BaseIntegrationTestsConfig>(TestUtilTypes.BaseIntegrationTestsConfig).to(IModelsClientsTestsConfig).inSingletonScope();
+  TestUtilBootstrapper.bind(container, path.join(__dirname, "..", "..", "..", ".env"));
+
   return container;
 }
