+++ tests/imodels-clients-tests/src/integration/common/index.ts
@@ -2,7 +2,6 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export * from "./IModelsClientsTestsConfig";
 export * from "./Constants";
 export * from "./GlobalSetup.test";
 export * from "./TestDiContainerProvider";
