+++ utils/imodels-client-test-utils/src/TestUtilTypes.ts
@@ -3,6 +3,5 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 export const TestUtilTypes = {
-  IModelsClientOptions: Symbol("IModelsClientOptions"),
-  BaseIntegrationTestsConfig: Symbol("BaseIntegrationTestsConfig")
+  IModelsClientOptions: Symbol("IModelsClientOptions")
 };
