+++ utils/imodels-client-test-utils/src/test-context-providers/auth/TestAuthorizationClient.ts
@@ -5,7 +5,7 @@
 import { ParsedUrlQuery } from "querystring";
 import { URLSearchParams, parse } from "url";
 import axios, { AxiosResponse } from "axios";
-import { inject, injectable } from "inversify";
+import { injectable } from "inversify";
 import * as puppeteer from "puppeteer";
 import { TestSetupError } from "../../CommonTestUtils";
 import { TestAuthorizationClientConfig } from "./TestAuthorizationClientConfigImpl";
@@ -38,7 +38,6 @@ export class TestAuthorizationClient {
   };
 
   constructor(
-    @inject(TestAuthorizationClientConfig)
     private readonly _authConfig: TestAuthorizationClientConfig
   ) { }
