+++ utils/imodels-client-test-utils/src/test-context-providers/auth/TestAuthorizationProviderConfig.ts
@@ -2,9 +2,8 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { inject, injectable } from "inversify";
-import { BaseIntegrationTestsConfig, TestUsersConfigValues } from "../../BaseIntegrationTestsConfig";
-import { TestUtilTypes } from "../../TestUtilTypes";
+import { injectable } from "inversify";
+import { IModelsClientsTestsConfig, TestUsersConfigValues } from "../../IModelsClientsTestsConfig";
 
 interface ApiScopes {
   iModels: string;
@@ -17,8 +16,7 @@ export class TestAuthorizationProviderConfig {
   public apiScopes: ApiScopes;
 
   constructor(
-    @inject(TestUtilTypes.BaseIntegrationTestsConfig)
-    config: BaseIntegrationTestsConfig
+    config: IModelsClientsTestsConfig
   ) {
     this.testUsers = config.testUsers;
     this.apiScopes = {
