+++ utils/imodels-client-test-utils/src/test-context-providers/imodel/TestModelsClientOptions.ts
@@ -2,18 +2,16 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { inject, injectable } from "inversify";
+import { injectable } from "inversify";
 import { ApiOptions, IModelsClientOptions } from "@itwin/imodels-client-authoring";
-import { BaseIntegrationTestsConfig } from "../../BaseIntegrationTestsConfig";
-import { TestUtilTypes } from "../../TestUtilTypes";
+import { IModelsClientsTestsConfig } from "../../IModelsClientsTestsConfig";
 
 @injectable()
 export class TestIModelsClientOptions implements IModelsClientOptions {
   public api: ApiOptions;
 
   constructor(
-    @inject(TestUtilTypes.BaseIntegrationTestsConfig)
-    config: BaseIntegrationTestsConfig
+    config: IModelsClientsTestsConfig
   ) {
     this.api = { baseUrl: config.apis.iModels.baseUrl };
   }
