+++ utils/imodels-client-test-utils/src/test-context-providers/project/ProjectsClient.ts
@@ -3,9 +3,10 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 import axios, { AxiosResponse } from "axios";
-import { inject, injectable } from "inversify";
+import { injectable } from "inversify";
 import { AuthorizationParam } from "@itwin/imodels-client-authoring";
 import { ProjectsClientConfig } from "./ProjectsClientConfig";
+import "reflect-metadata";
 
 interface Project {
   id: string;
@@ -22,7 +23,6 @@ interface ProjectResponse {
 @injectable()
 export class ProjectsClient {
   constructor(
-    @inject(ProjectsClientConfig)
     private _config: ProjectsClientConfig
   ) { }
