+++ utils/imodels-client-test-utils/src/test-context-providers/project/ProjectsClientConfig.ts
@@ -2,17 +2,15 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { inject, injectable } from "inversify";
-import { BaseIntegrationTestsConfig } from "../../BaseIntegrationTestsConfig";
-import { TestUtilTypes } from "../../TestUtilTypes";
+import { injectable } from "inversify";
+import { IModelsClientsTestsConfig } from "../../IModelsClientsTestsConfig";
 
 @injectable()
 export class ProjectsClientConfig {
   public baseUrl: string;
 
   constructor(
-    @inject(TestUtilTypes.BaseIntegrationTestsConfig)
-    config: BaseIntegrationTestsConfig
+    config: IModelsClientsTestsConfig
   ) {
     this.baseUrl = config.apis.projects.baseUrl;
   }
