+++ utils/imodels-client-test-utils/src/test-context-providers/project/TestProjectProviderConfig.ts
@@ -2,15 +2,15 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { inject, injectable } from "inversify";
-import { BaseIntegrationTestsConfig } from "../../BaseIntegrationTestsConfig";
-import { TestUtilTypes } from "../../TestUtilTypes";
+import { injectable } from "inversify";
+import { IModelsClientsTestsConfig } from "../../IModelsClientsTestsConfig";
+import "reflect-metadata";
 
 @injectable()
 export class TestProjectProviderConfig {
   public testProjectName: string;
 
-  constructor(@inject(TestUtilTypes.BaseIntegrationTestsConfig) config: BaseIntegrationTestsConfig) {
+  constructor(config: IModelsClientsTestsConfig) {
     this.testProjectName = config.testProjectName;
   }
 }
