+++ utils/imodels-client-test-utils/src/test-imodel-group/TestIModelGroupFactory.ts
@@ -2,18 +2,16 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { inject, injectable } from "inversify";
+import { injectable } from "inversify";
 import { TestAuthorizationProvider, TestIModelsClient, TestProjectProvider } from "../test-context-providers";
 import { TestIModelGroup } from "./TestIModelGroup";
+import "reflect-metadata";
 
 @injectable()
 export class TestIModelGroupFactory {
   constructor(
-    @inject(TestIModelsClient)
     private readonly _iModelsClient: TestIModelsClient,
-    @inject(TestAuthorizationProvider)
     private readonly _testAuthorizationProvider: TestAuthorizationProvider,
-    @inject(TestProjectProvider)
     private readonly _testProjectProvider: TestProjectProvider
   ) { }
